Move Player CPE stuff into ClassicProtocol and Player.Networking.cs
